Move unlink_try_hard and dependents into mkosi.remove
Similar to mkosi.install but for removing things.
