+++ mkosi/__init__.py
@@ -113,6 +113,7 @@ from mkosi.install import (
 )
 from mkosi.manifest import Manifest
 from mkosi.mounts import mount, mount_api_vfs, mount_bind, mount_overlay, mount_tmpfs
+from mkosi.remove import unlink_try_hard
 from mkosi.syscall import blkpg_add_partition, blkpg_del_partition
 
 complete_step = MkosiPrinter.complete_step
@@ -559,27 +560,6 @@ def setup_workspace(config: MkosiConfig) -> TempDir:
 def btrfs_subvol_create(path: Path, mode: int = 0o755) -> None:
     with set_umask(~mode & 0o7777):
         run(["btrfs", "subvol", "create", path])
-
-
-def btrfs_subvol_delete(path: Path) -> None:
-    # Extract the path of the subvolume relative to the filesystem
-    c = run(["btrfs", "subvol", "show", path],
-            stdout=subprocess.PIPE, stderr=subprocess.DEVNULL, text=True)
-    subvol_path = c.stdout.splitlines()[0]
-    # Make the subvolume RW again if it was set RO by btrfs_subvol_delete
-    run(["btrfs", "property", "set", path, "ro", "false"])
-    # Recursively delete the direct children of the subvolume
-    c = run(["btrfs", "subvol", "list", "-o", path],
-            stdout=subprocess.PIPE, stderr=subprocess.DEVNULL, text=True)
-    for line in c.stdout.splitlines():
-        if not line:
-            continue
-        child_subvol_path = line.split(" ", 8)[-1]
-        child_path = path / cast(str, os.path.relpath(child_subvol_path, subvol_path))
-        btrfs_subvol_delete(child_path)
-    # Delete the subvolume now that all its descendants have been deleted
-    run(["btrfs", "subvol", "delete", path],
-        stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
 
 
 def btrfs_subvol_make_ro(path: Path, b: bool = True) -> None:
@@ -5260,29 +5240,6 @@ def detect_distribution() -> Tuple[Optional[Distribution], Optional[str]]:
         version_id = version_codename or extracted_codename
 
     return d, version_id
-
-
-def unlink_try_hard(path: Optional[PathString]) -> None:
-    if path is None:
-        return
-
-    path = Path(path)
-    try:
-        path.unlink()
-        return
-    except FileNotFoundError:
-        return
-    except Exception:
-        pass
-
-    if shutil.which("btrfs"):
-        try:
-            btrfs_subvol_delete(path)
-            return
-        except Exception:
-            pass
-
-    shutil.rmtree(path)
 
 
 def remove_glob(*patterns: PathString) -> None:
